termine las entities y termine documentacion en entiites.

+++ tiendaApp/src/main/java/edu/unimag/entities/OrdenCompra.java
@@ -2,8 +2,6 @@ package edu.unimag.entities;
 
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -38,7 +36,7 @@ public class OrdenCompra {
 
     @Schema(description = "Total de la compra", example = "299000")
     @Column(nullable = false)
-    private BigDecimal total;
+    private Double total;
 
     // Relación con proveedor 
     @Schema(description = "Proveedor asociado a la orden")
